+++ packages/storybook/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @stcland/storybook
 
+## 0.0.14
+
+### Patch Changes
+
+- Added indeterminate prop to Checkbox.
+- Updated dependencies
+  - @stcland/components@0.0.10
+
+## 0.0.13
+
+### Patch Changes
+
+- Added checkbox independent from form to STCLandia.
+- Updated dependencies
+  - @stcland/components@0.0.9
+
 ## 0.0.12
 
 ### Patch Changes